+++ src/main/java/com/sap/oss/phosphor/fosstars/tool/format/MarkdownElement.java
@@ -1,0 +1,6 @@
+package com.sap.oss.phosphor.fosstars.tool.format;
+
+public interface MarkdownElement {
+
+  String make();
+}
