+++ src/main/java/com/sap/oss/phosphor/fosstars/tool/format/OrderedMarkdownList.java
@@ -1,0 +1,10 @@
+package com.sap.oss.phosphor.fosstars.tool.format;
+
+import java.util.List;
+
+public class OrderedMarkdownList extends MarkdownList {
+
+  OrderedMarkdownList(List<MarkdownElement> elements) {
+    super(elements, "1.  ");
+  }
+}
