+++ src/main/java/com/sap/oss/phosphor/fosstars/tool/format/UnorderedMarkdownList.java
@@ -1,0 +1,10 @@
+package com.sap.oss.phosphor.fosstars.tool.format;
+
+import java.util.List;
+
+public class UnorderedMarkdownList extends MarkdownList {
+
+  UnorderedMarkdownList(List<MarkdownElement> elements) {
+    super(elements, "*  ");
+  }
+}
